Update nicknames to only use 2 bytes.
If every Block had an unique miner, it would take 1.25 years to fill the 2-byte space.

+++ NimTests/DatabaseTests/MeritTests/BlockchainTests/BDBTest.nim
@@ -127,7 +127,7 @@ proc test*() =
                 blockchain.tip.header.hash,
                 contents.hash,
                 verifiers,
-                uint32(miner),
+                uint16(miner),
                 miners[miner],
                 transactions,
                 elements
