+++ src/Database/Consensus/Elements/objects/VerificationObj.nim
@@ -18,7 +18,7 @@ import finals
 finalsd:
     type
         Verification* = ref object of Element
-            holder*: uint32
+            holder*: uint16
             hash* {.final.}: Hash[384]
 
         SignedVerification* = ref object of Verification
